requirements.txt
numpy
pandas
bs4
python-dateutil 
requests
flair
plotly
tqdm